requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-magic
pdfminer.six
pillow
pytesseract
beautifulsoup4
sqlalchemy>=2.0
psycopg2-binary
alembic

